indexer/requirements.txt
faiss-cpu==1.7.4
numpy==1.26.4
openai==1.10.0
watchfiles==0.21.0
httpx==0.27.2
sentence-transformers==2.2.2
torch==2.2.1
pytest==7.4.4
huggingface_hub==0.24.4 